backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
aioredis
pika
python-dotenv
alembic
python-jose[cryptography]
passlib
bcrypt
email-validator
pydantic-settings
cloudinary
python-multipart

# Testing
pytest
pytest-asyncio
pytest-cov
httpx



